Add factory funciton. Since there was a circular reference, the main skate function had to be moved from index to it's own api file.
