+++ test/unit.js
@@ -2,6 +2,7 @@ import './boot';
 import './unit/index';
 import './unit/api/create';
 import './unit/api/emit';
+import './unit/api/factory';
 import './unit/api/fragment';
 import './unit/api/init';
 import './unit/api/properties';
